Extract PyZX code into separate file
This is because we want to treat PyZX as an optional dependency.
Since OpenGraph is meant to be a standardised interface for many
projects to use, it shouldn't include any optional dependencies
